Improved overlap detection
Overlapping detection doesn't have to be based on pre-defined loop length.
Since SJCam records the end of recording time as "creation_time" in the metadata
and the length of the video can also be found out using ffprobe, it's just simple
math to find out if two files overlaps.
Though during development I found some interesting behavior on behalf of SJCam's
timestamping. See the added comments in the code.