GPS related improvements
The major improvement of this commit is, that the sjscan now looks into the given
directory for gpx files. Assuming the same file naming standard gpslog uses.
The main gpx file is copied following the Session file's naming, but only the
entries kept where the timestamp is between the session's time boundary.
(Plus-Minus 30 seconds on the ends. By default)

Minor improvements:
- sjdemux has a more sensible default for file name prefix.
- vedit.sh has the sjrender function documented
- README updates

Bug fixes:
- A few typo here and there in sjdemux looking for defaults in environment variables