followup for new avidemux version and some minor fixes
With the quaterly pkg update Avidemux 2.8.1 was installed.
This means, another few stub method needed to be added to the sjdemux
script.

Also, going back to calculate the length of a video based on the video
stream's length, not the "format" length.
Some of my SJCam recorded files were 601 seconds long in the video
stream, but according to the metadata, the format and the audio stream's
length was 601.088 second long.

In the Video generation process, I had to simplify the ffmpeg process,
because when I added the audio and the video overlay rendering at the
same command, it somehow always made the sound to skew on the long run.