gpdemux and vedit fixes and improvements
When a gopro session has been split into mutliple files, the start time is the same for all of the mp4.
But in the calculations, I have to count in the previous segment's length as well.
This commit intents to solve that issue.

I also made some improvements in the vedit.sh shell helper collection:
Overlay extension can be anything. Either png or tiff.
On the comments of gpx2video, someone mentioned, that exporting the overlay images are much faster into tiff, than png.
This change intend to manage that.