Blank file support
Minor improvments:
- sjmpv can now play more than just one file in the main player
- sjstgen generates 25fps kdenlive file
- new examples in the examples subdir

Major improvement: Blank file support!
See the `examples/blackgen.sh` file how to generate a 2 hour long 4k resolution video which is completely blank all the time.
If you pass this to gpdemux it will create ffmpeg commands to stream copy from this small file: the temporary files generated to use as reference for the gpx2video is much much smaller. File generation finishes pretty quick, since there's not too much io to do. And vedit.sh is aware of this improvement, so if it finds an .input.txt file next to the .mp4 file with the same name, it will use the .input.txt file with the concat filter as input, so no need for that slices to cut from the original camera recorded material. The ffmpeg will use the "source of truth" directly.