Dir structure reorg + new examples
I did some rework on the dir structore. Earlier I had to create a WK.MMDD
directory for the py project files and the overlay files, and a WKO.MMDD
directory for the rendering output.

New structure: In the workdir there are subdirectories. Once you found out
the difference between your camera's clock and GPS, you just render the
"blank" files, and generate the overlays.
The new gpgentimediff function will generate you a timediff.sh where the
script generates the output into the fsrc subdir.
Also, there's a new function for the overlay image rendering, which takes
this new structure into consideration.
Now all images will be generated and goes into the fovl.
In case you do any edit on the overlays, you don't have to regenerate the
overlay images. You have them all. The renderall will call another
function called gentemp, to setup a tempdir for the rendering, and will do
symlinks for the overlay images.
For the gentemp() to be able to do that, the gpdemux generates not just
the ffmpeg commands, but also a helper-file with the .dsh extension.
It's syntactically a shell file, what you can source in, and sets a few
variables:
l_blank -> will be yes, if the generated mp4 file is just a placeholder
  with the right creation_time set, but the .input.txt file should be used
l_duration -> The length of this segment in seconds
l_sdiff -> Where this segment starts compared to the uncut camera segment.

Also, I've added an example for a fadein effect.
The example file make a 1 send long fadein, so the last 1 second of the
"current" segment will be hidden by the end of that segment starting
from t-1, and the new segment will be there.
The example has all the "neccesary math" to calculate the neccesary
paddings on both sides. (Original audio needs an "apad", new audio needs
an "adelay". Than the two needs to be "amerge"d.
Second video stream needs to be generated (skipping the frames for
speedup), adding the overlay, fadein, padding.
Than this needs to be merge with the original video with another use of
the overlay filter.
Also, since the next segment is already merged into the "current" one,
the next segment needs to be not rendered, so I have an example to
direct that to /dev/null.

When gpdemux is run in diff mode, it will give you a command example
what can go to the work.sh almost directly. (the ...u.py needs to be
...n.py)
But the gptelltimediffs will take care of that.

Also, this makes the usage a few steps easier:
1. gpscan
2. gpgentimesh
3. vi timediff.sh with the proper starting timediff
4. Run the . timediff.sh
5. run the gptelltimediffs and it will generate the work.sh
6. . work.sh
7. In gpx2video's build dir: sjrenderimg 4kpub p
   Also, it now only needs just one symlink, the wk pointing to the
   project dir
8. renderall p
9. concatall p
10. do the music stuff
	modified:   vedit.sh